separate recovery from boostrap
